requirements.txt
kroger-api
networkx
matplotlib
python-dotenv
ollama
